p3/shelfscout_backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.24
python-multipart>=0.0.9
pydantic>=2.5
pillow>=10.0
numpy>=1.24

# Install torch/torchvision according to your CPU/CUDA setup:
# https://pytorch.org/get-started/locally/
torch
torchvision
